feat: implementando regras de negócio

+++ backend/src/modules/positions/positions.service.ts
@@ -1,3 +1,5 @@
+import { ConflictError } from 'src/common/errors/types/ConflictError';
+import { NotFoundError } from 'src/common/errors/types/NotFoundError';
 import { Injectable } from '@nestjs/common';
 import { CreatePositionDto } from './dto/create-position.dto';
 import { UpdatePositionDto } from './dto/update-position.dto';
@@ -7,8 +9,12 @@ import { PositionRepository } from './repositories/position.repository';
 export class PositionsService {
   constructor(private readonly repository: PositionRepository) {}
 
-  async create(position: CreatePositionDto) {
-    return await this.repository.create(position);
+  async create(createPositionDto: CreatePositionDto) {
+    const position = createPositionDto.position;
+    const verifyPosition = await this.repository.verifyExisteField(position);
+    if (verifyPosition) throw new ConflictError('position already exists');
+
+    return await this.repository.create(createPositionDto);
   }
 
   async findAll() {
@@ -16,14 +22,20 @@ export class PositionsService {
   }
 
   async findOne(id: number) {
+    const position = await this.repository.findOne(id);
+    if (!position) throw new NotFoundError(`Position ${id} is not found`);
     return await this.repository.findOne(id);
   }
 
   async update(id: number, updatePositionDto: UpdatePositionDto) {
+    const position = await this.repository.findOne(id);
+    if (!position) throw new NotFoundError(`Position ${id} is not found`);
     return await this.repository.update(id, updatePositionDto);
   }
 
   async remove(id: number) {
+    const position = await this.repository.findOne(id);
+    if (!position) throw new NotFoundError(`Position ${id} is not found`);
     return await this.repository.remove(id);
   }
 }
